Add Bitcoin block hash to proof-of-freshness
Rational as Joanna Rutkowska put it:

1. Simple
2. Doesn't trust external organizations
3. Doesn't require working RSS feeds to be verified

Blockchain.info's API is used, with the 10th most-recent block hash used rather
than the most recent to (drastically) reduce the risk of a re-org invalidating
the hash used. That said, if that does happen (or blockchain.info screws up)
it's not the end of the world: we still have the RSS feeds as backup.