Update QSB template
This commit assumes that QubesOS/qubes-issues#6692 has been implemented.

Separates the initial section into "User action required" and
"Patching." "User action required" is for any special intructions that
users must follow beyond updating the system normally. For most QSBs, it
is not expected that there will be any such special instructions, so the
default text provided here should suffice.

The "Patching" section is for advanced users who are interested in
technical details about the packages or in testing them while they're in
security-testing.