Canary-029: Reword SHA-1 explaination
Exploitation of a SHA-1 collision would allow a malicious actor to make
us sign a key we did not intend to sign. Since the QMSK only signs code
signing keys (and for historic reasons security@) this is a very
theoretic concerns since such an attacker would have already gained
direct access to a code signing key. Also in other PGP use cases this is
more relevant. Therefore "does not pose any security risk" is a bit too
strong statement, so I weakened the statement a bit.

Additionaly mention the main reason for doing this now: Allow disabling
of SHA-1 support in tooling.